We are now making sure that we work with the 1.6 versions of Angular.
Angular used to provide success and error functions on the request object and we used those to respond to request and do token handling.  However, as of Angular version 1.6 they have removed these two functions and instead require you to use the then function which has two callbacks, one for success and one for error.

We are now using the new function so we can support Angular 1.6 in Gromit.